pyproject.toml
[tool.poetry]
name = "retrofun"
version = "0.1.0"
description = "Code From 'SQLAlchemy 2 in Practice' by Miguel Grinberg"
authors = ["Grant Ramsay <seapagan@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.18"
psycopg2 = "^2.9.6"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
ipython = "^8.14.0"
bandit = { extras = ["toml"], version = "^1.7.5" }
black = "^23.3.0"
flake8 = "^6.0.0"
flake8-docstrings = "^1.7.0"
flake8-pyproject = "^1.2.3"
flake8-pytest-style = "^1.7.2"
flake8-type-checking = "^2.4.0"
isort = "^5.12.0"
pylint = "^2.17.2"
pep8-naming = "^0.13.3"
pre-commit = "^3.3.3"
pydocstyle = "^6.3.0"
mypy = "^1.4.1"
tryceratops = "^2.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# configure assorted tools and linters
[tool.isort]
profile = "black"

[tool.black]
line-length = 80
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | venv
  | \.venv
  | _build
  | build
  | dist
)/
'''

[tool.flake8]
max-line-length = 80
exclude = ["__init__.py", ".git"]
extend-ignore=["E203", "W503"]
extend-select=["TC", "TC1", "TRY"]
docstring-convention = "google"
type-checking-pydantic-enabled = true

[tool.bandit]
exclude_dirs = []

[tool.bandit.assert_used]
skips = ['*_test.py', '*/test_*.py']

[tool.pydocstyle]
add-ignore = ["D104"]

[tool.pytest.ini_options]
filterwarnings = []
